Fix grid rendering to clip to the actual visible size, not the maximum scrollable extent, use doubles where appropriate & increases the label step-size so you don't end up with labels that overlap using the worst case label size. The latter causes the grid to animate but that's because the grid lines should really be aligned to time increments that make sense rather than being done so arbitrarily.

+++ src/view/src/rocprofvis_grid.h
@@ -17,7 +17,7 @@ public:
     ~Grid();
 
     void   RenderGrid(double min_x, double max_x, double movement, float zoom,
-                      float scale_x, float v_max_x, float v_min_x, int grid_size,
+                      float scale_x, double v_max_x, double v_min_x, int grid_size,
                       int sidebar_size);
     double  GetCursorPosition(float mouse_position);
     double GetViewportStartPosition();
